Refactor: Extract choosing and clicking episode from watch_media
