+++ src/cli/mod.rs
@@ -1,5 +1,4 @@
 pub mod choose_episode;
-pub mod choose_episode_with_images;
 pub mod choose_lang;
 pub mod choose_media;
 pub mod choose_media_with_images;
